Code review: Fix client null deref crashes in game table refresh functions and animation callbacks (add null guards for game/hand/session in refreshSet/refreshButton/refreshAction/refreshAll/refreshChangePlayer/refreshPot/refreshGameLabels/disableMyButtons/mySet/postRiverRunAnimation2-6/showHoleCards and all *Animation*Action timer callbacks that fire after game ends), fix ConfigFile writeBuffer/updateConfig using defaultValue as XML element name instead of name for list sub-elements (same bug as previously fixed in CleanerConfig), fix getMyCallAmount() integer underflow when humanPlayer.getMySet() > tempHighestSet (add std::max(0,...)), fix mySet() calling setMySet() twice on all-in path and once before all-in check (compute betValue first, set once in correct branch), fix chattools nick highlight matching on HTML-escaped text instead of original message (check nick before toHtmlEscaped), fix guilog null deref crashes on corrupted .pdb log files (add null checks for sqlite3 result columns), fix PlayerData::IsPlayerAllowedToJoinCreateLimitRank integer overflow for large periodMinutes (cast to long long before multiply), fix mytimeoutlabel unsigned underflow when realTimerValue < 3000 or timeOutValue <= 3 (add guards), fix AndroidAudio::reallyPlaySound inserting null entries into mSounds map via operator[] (use value() instead), fix CleanerConfig::readConfigInt silent parse failure returning 0 for missing/invalid variables (add found check and parse error logging)

+++ src/playerdata.cpp
@@ -291,7 +291,7 @@ PlayerData::IsPlayerAllowedToJoinCreateLimitRank(const std::string &num, const s
 		return false;
 	}
 
-	long then = static_cast<long>(time(nullptr)) - static_cast<long>(periodMinutes) * 60;
+	long long then = static_cast<long long>(time(nullptr)) - static_cast<long long>(periodMinutes) * 60LL;
 
 	m_last_games.erase(
 		std::remove_if(m_last_games.begin(), m_last_games.end(),
